fixed ui + deleted sidemenuscript
