+++ Assets/Images/close-button-800px.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e5c782fb3c518584a9d6a1a7f1626b4a
+guid: b7f2415f3dee8614cabd215fdb9a7989
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
