+++ Assets/Scenes/MainScene.unity
@@ -1952,7 +1952,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 310932521}
-  - component: {fileID: 310932522}
   m_Layer: 5
   m_Name: SideMenuExistingLocation
   m_TagString: Untagged
@@ -1983,19 +1982,6 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 270.2927, y: 468.8}
   m_Pivot: {x: 0.5, y: 0.5}
---- !u!114 &310932522
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 310932520}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e58ac720b7f6e9b4994e03ec8333e4fa, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  showMorePanel: {fileID: 0}
 --- !u!1 &325367288
 GameObject:
   m_ObjectHideFlags: 0
@@ -2323,8 +2309,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 310932522}
-        m_TargetAssemblyTypeName: SideMenuExistLocation, Scripts
+      - m_Target: {fileID: 872695253}
+        m_TargetAssemblyTypeName: SpatialNotes.MapInteract, Scripts
         m_MethodName: DeleteLocation
         m_Mode: 1
         m_Arguments:
@@ -7482,8 +7468,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 310932522}
-        m_TargetAssemblyTypeName: SideMenuExistLocation, Scripts
+      - m_Target: {fileID: 872695253}
+        m_TargetAssemblyTypeName: SpatialNotes.MapInteract, Scripts
         m_MethodName: EditLocation
         m_Mode: 1
         m_Arguments:
@@ -8239,9 +8225,9 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 310932522}
-        m_TargetAssemblyTypeName: SideMenuExistLocation, Scripts
-        m_MethodName: Close
+      - m_Target: {fileID: 872695253}
+        m_TargetAssemblyTypeName: SpatialNotes.MapInteract, Scripts
+        m_MethodName: CloseSideExistingMenu
         m_Mode: 1
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -9686,8 +9672,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 310932522}
-        m_TargetAssemblyTypeName: SideMenuExistLocation, Scripts
+      - m_Target: {fileID: 872695253}
+        m_TargetAssemblyTypeName: SpatialNotes.MapInteract, Scripts
         m_MethodName: CloseShowMore
         m_Mode: 1
         m_Arguments:
@@ -9819,8 +9805,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 310932522}
-        m_TargetAssemblyTypeName: SideMenuExistLocation, Scripts
+      - m_Target: {fileID: 872695253}
+        m_TargetAssemblyTypeName: SpatialNotes.MapInteract, Scripts
         m_MethodName: Toggle
         m_Mode: 1
         m_Arguments:
@@ -10774,8 +10760,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 310932522}
-        m_TargetAssemblyTypeName: SideMenuExistLocation, Scripts
+      - m_Target: {fileID: 872695253}
+        m_TargetAssemblyTypeName: SpatialNotes.MapInteract, Scripts
         m_MethodName: FavoriteLocation
         m_Mode: 1
         m_Arguments:
