--- Assets/Scripts/SideMenuExistLocation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e58ac720b7f6e9b4994e03ec8333e4fa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
