+++ Assets/StreamingAssets/Maps/testmap/assetsassetshell.png.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: acee3c3075a339240b2d9bc8c7d5c852
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
