+++ Assets/StreamingAssets/Maps/testmap/notes/location.json
@@ -1,1 +1,1 @@
-{"keys":["36.102,58.56,0","29.426,41.051,0","54.268,50,0","45.433,47.91,0","-1.206691,11.88128,0","-24.84982,2.626936,0","30.12934,-14.84006,0","37.07116,-57.46016,0","11.10988,5.887301,0","-22.55587,23.20563,0","-83.63748,14.01534,0","-18.36816,-34.28666,0","-47.80359,-0.1856483,0","-60.42745,2.784674,0"],"values":[{"locationName":"dsa","description":"ads","coordinate":{"x":36.10200119018555,"y":58.560001373291019,"z":0.0},"imagePath":""},{"locationName":"dsa","description":"ads","coordinate":{"x":29.426000595092775,"y":41.05099868774414,"z":0.0},"imagePath":""},{"locationName":"dsa","description":"ads","coordinate":{"x":54.268001556396487,"y":50.0,"z":0.0},"imagePath":""},{"locationName":"test1","description":"a","coordinate":{"x":45.43299865722656,"y":47.90999984741211,"z":0.0},"imagePath":""},{"locationName":"test2","description":"t2","coordinate":{"x":-1.206691026687622,"y":11.881279945373536,"z":0.0},"imagePath":""},{"locationName":"test3","description":"as","coordinate":{"x":-24.84981918334961,"y":2.6269359588623049,"z":0.0},"imagePath":""},{"locationName":"philly ","description":"skull emoji","coordinate":{"x":30.12933921813965,"y":-14.840060234069825,"z":0.0},"imagePath":""},{"locationName":"panama","description":"","coordinate":{"x":37.07115936279297,"y":-57.46015930175781,"z":0.0},"imagePath":""},{"locationName":"Russia","description":"","coordinate":{"x":11.109880447387696,"y":5.887300968170166,"z":0.0},"imagePath":""},{"locationName":"que","description":"why es","coordinate":{"x":-22.555870056152345,"y":23.205629348754884,"z":0.0},"imagePath":""},{"locationName":"biden","description":"blast","coordinate":{"x":-83.63748168945313,"y":14.015339851379395,"z":0.0},"imagePath":"C:\\Users\\Rex Wang\\Downloads\\bidenblast.png"},{"locationName":"literal hell","description":"penn state or something","coordinate":{"x":-18.368160247802736,"y":-34.286659240722659,"z":0.0},"imagePath":"C:\\Users\\Rex Wang\\Downloads\\hell.png"},{"locationName":"basically hell","description":"kevin's house","coordinate":{"x":-47.8035888671875,"y":-0.18564830720424653,"z":0.0},"imagePath":"E:/Purdue/Spring 2024/ECE 49595/Assignments/spatialnotes/Assets/StreamingAssets/Maps/assetshell.png"},{"locationName":"hgfh","description":"s","coordinate":{"x":0.0,"y":0.0,"z":0.0},"imagePath":"E:/Purdue/Spring 2024/ECE 49595/Assignments/spatialnotes/Assets/StreamingAssets/Maps/testmap/assets1200px-Utah_teapot_(solid).stl.png"}]}
+{"keys":["36.102,58.56,0","29.426,41.051,0","54.268,50,0","45.433,47.91,0","-1.206691,11.88128,0","-24.84982,2.626936,0","30.12934,-14.84006,0","37.07116,-57.46016,0","11.10988,5.887301,0","-22.55587,23.20563,0","-83.63748,14.01534,0","-18.36816,-34.28666,0","-47.80359,-0.1856483,0","-60.42745,2.784674,0","-32.1727,-15.12465,0","-36.27737,-19.95352,0","-48.73219,-11.73093,0","-69.01136,-27.98271,0"],"values":[{"locationName":"dsa","description":"ads","coordinate":{"x":36.10200119018555,"y":58.560001373291019,"z":0.0},"imagePath":""},{"locationName":"dsa","description":"ads","coordinate":{"x":29.426000595092775,"y":41.05099868774414,"z":0.0},"imagePath":""},{"locationName":"dsa","description":"ads","coordinate":{"x":54.268001556396487,"y":50.0,"z":0.0},"imagePath":""},{"locationName":"test1","description":"a","coordinate":{"x":45.43299865722656,"y":47.90999984741211,"z":0.0},"imagePath":""},{"locationName":"test2","description":"t2","coordinate":{"x":-1.206691026687622,"y":11.881279945373536,"z":0.0},"imagePath":""},{"locationName":"test3","description":"as","coordinate":{"x":-24.84981918334961,"y":2.6269359588623049,"z":0.0},"imagePath":""},{"locationName":"philly ","description":"skull emoji","coordinate":{"x":30.12933921813965,"y":-14.840060234069825,"z":0.0},"imagePath":""},{"locationName":"panama","description":"","coordinate":{"x":37.07115936279297,"y":-57.46015930175781,"z":0.0},"imagePath":""},{"locationName":"Russia","description":"","coordinate":{"x":11.109880447387696,"y":5.887300968170166,"z":0.0},"imagePath":""},{"locationName":"que","description":"why es","coordinate":{"x":-22.555870056152345,"y":23.205629348754884,"z":0.0},"imagePath":""},{"locationName":"biden","description":"blast","coordinate":{"x":-83.63748168945313,"y":14.015339851379395,"z":0.0},"imagePath":"C:\\Users\\Rex Wang\\Downloads\\bidenblast.png"},{"locationName":"literal hell","description":"penn state or something","coordinate":{"x":-18.368160247802736,"y":-34.286659240722659,"z":0.0},"imagePath":"C:\\Users\\Rex Wang\\Downloads\\hell.png"},{"locationName":"basically hell","description":"kevin's house","coordinate":{"x":-47.8035888671875,"y":-0.18564830720424653,"z":0.0},"imagePath":"E:/Purdue/Spring 2024/ECE 49595/Assignments/spatialnotes/Assets/StreamingAssets/Maps/assetshell.png"},{"locationName":"hgfh","description":"s","coordinate":{"x":-60.42744827270508,"y":2.7846739292144777,"z":0.0},"imagePath":"E:/Purdue/Spring 2024/ECE 49595/Assignments/spatialnotes/Assets/StreamingAssets/Maps/testmap/assets1200px-Utah_teapot_(solid).stl.png"},{"locationName":"qweqweqwe","description":"qweqweqwe","coordinate":{"x":-32.172698974609378,"y":-15.124650001525879,"z":0.0},"imagePath":"C:/Users/klin0/SpatialNotes/Assets/StreamingAssets/Maps/testmap/assetsassetshell.png"},{"locationName":"What the","description":"","coordinate":{"x":-36.27737045288086,"y":-19.953519821166993,"z":0.0},"imagePath":""},{"locationName":"What the 2","description":"Kevin's mind","coordinate":{"x":-48.7321891784668,"y":-11.73093032836914,"z":0.0},"imagePath":"C:/Users/klin0/SpatialNotes/Assets/StreamingAssets/Maps/testmap/assetsassetshell.png"},{"locationName":"I want to see if this works","description":"Lets see...","coordinate":{"x":0.0,"y":0.0,"z":0.0},"imagePath":"C:/Users/klin0/SpatialNotes/Assets/StreamingAssets/Maps/testmap/assetsassetshell.png"}]}
